ROO-2337: Post 1.1.3.RELEASE code refactor and clean up - remove unnecessary casts

+++ support/src/main/java/org/springframework/roo/support/util/Base64.java
@@ -1199,7 +1199,7 @@ public class Base64
             else {
                 // There's a bad input character in the Base64 stream.
                 throw new java.io.IOException( String.format(
-                "Bad Base64 input character decimal %d in array position %d", ((int)source[i])&0xFF, i ) );
+                "Bad Base64 input character decimal %d in array position %d", source[i] & 0xFF, i ) );
             }   // end else: 
         }   // each input character
                                    
@@ -1260,7 +1260,7 @@ public class Base64
         boolean dontGunzip = (options & DONT_GUNZIP) != 0;
         if( (bytes != null) && (bytes.length >= 4) && (!dontGunzip) ) {
             
-            int head = ((int)bytes[0] & 0xff) | ((bytes[1] << 8) & 0xff00);
+            int head = (bytes[0] & 0xff) | ((bytes[1] << 8) & 0xff00);
             if( java.util.zip.GZIPInputStream.GZIP_MAGIC == head )  {
                 java.io.ByteArrayInputStream  bais = null;
                 java.util.zip.GZIPInputStream gzis = null;
